# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sehhat Foundation Online Exam System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface StudentLoginRequest {
  studentId: number;
  programId: number;
  yearId: number;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  faculty: "faculty",
  student: "student",
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  programId?: number | null;
  yearId?: number | null;
  rollNumber?: string | null;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  faculty: "faculty",
  student: "student",
} as const;

export interface CreateUserRequest {
  username: string;
  name: string;
  password: string;
  role: CreateUserRequestRole;
  programId?: number | null;
  yearId?: number | null;
  rollNumber?: string | null;
}

export interface UpdateUserRequest {
  username?: string;
  name?: string;
  password?: string | null;
  programId?: number | null;
  yearId?: number | null;
  rollNumber?: string | null;
}

export interface StudentInfo {
  id: number;
  name: string;
  rollNumber?: string | null;
}

export interface Year {
  id: number;
  label: string;
  value: number;
}

export interface CreateYearRequest {
  label: string;
  value: number;
}

export interface Program {
  id: number;
  name: string;
  code: string;
  description?: string | null;
}

export interface CreateProgramRequest {
  name: string;
  code: string;
  description?: string | null;
}

export interface Semester {
  id: number;
  number: number;
  label: string;
}

export interface Subject {
  id: number;
  name: string;
  code: string;
  programId: number;
  semesterId: number;
  programName?: string | null;
  semesterLabel?: string | null;
}

export interface CreateSubjectRequest {
  name: string;
  code: string;
  programId: number;
  semesterId: number;
}

export interface Exam {
  id: number;
  title: string;
  subjectId: number;
  subjectName?: string | null;
  programId: number;
  programName?: string | null;
  semesterId: number;
  yearId?: number | null;
  durationMinutes: number;
  totalQuestions: number;
  isActive: boolean;
  mcqCount?: number | null;
  createdAt: string;
}

export interface CreateExamRequest {
  title: string;
  subjectId: number;
  programId: number;
  semesterId: number;
  yearId?: number | null;
  durationMinutes: number;
  totalQuestions: number;
  isActive: boolean;
}

export type McqCorrectOption =
  (typeof McqCorrectOption)[keyof typeof McqCorrectOption];

export const McqCorrectOption = {
  A: "A",
  B: "B",
  C: "C",
  D: "D",
} as const;

export interface Mcq {
  id: number;
  examId: number;
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctOption: McqCorrectOption;
  questionNumber: number;
}

export type McqInputCorrectOption =
  (typeof McqInputCorrectOption)[keyof typeof McqInputCorrectOption];

export const McqInputCorrectOption = {
  A: "A",
  B: "B",
  C: "C",
  D: "D",
} as const;

export interface McqInput {
  questionText: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctOption: McqInputCorrectOption;
  questionNumber: number;
}

export interface BulkMcqRequest {
  mcqs: McqInput[];
}

export interface ExamWithMcqs {
  exam: Exam;
  mcqs: Mcq[];
}

export type ShuffledQuestionShuffledOptionsItem = {
  label: string;
  text: string;
};

export interface ShuffledQuestion {
  id: number;
  questionText: string;
  shuffledOptions: ShuffledQuestionShuffledOptionsItem[];
  displayNumber: number;
}

export type ExamSessionExistingAnswers = { [key: string]: string };

export interface ExamSession {
  attemptId: number;
  examId: number;
  questions: ShuffledQuestion[];
  durationMinutes: number;
  startedAt: string;
  existingAnswers: ExamSessionExistingAnswers;
}

export interface SaveAnswerRequest {
  attemptId: number;
  questionId: number;
  selectedOption: string;
}

export type SubmitExamRequestAnswers = { [key: string]: string };

export type SubmitExamRequestSubmissionReason =
  (typeof SubmitExamRequestSubmissionReason)[keyof typeof SubmitExamRequestSubmissionReason];

export const SubmitExamRequestSubmissionReason = {
  manual: "manual",
  timeout: "timeout",
  tab_switch: "tab_switch",
} as const;

export interface SubmitExamRequest {
  attemptId: number;
  answers: SubmitExamRequestAnswers;
  submissionReason: SubmitExamRequestSubmissionReason;
}

export interface ExamResult {
  id: number;
  studentId: number;
  studentName?: string | null;
  rollNumber?: string | null;
  examId: number;
  examTitle?: string | null;
  subjectName?: string | null;
  programName?: string | null;
  score: number;
  totalQuestions: number;
  percentage: number;
  submissionReason: string;
  submittedAt: string;
}

export interface LiveExamAttempt {
  attemptId: number;
  studentName?: string | null;
  rollNumber?: string | null;
  examTitle?: string | null;
  startedAt: string;
  answeredCount: number;
  totalQuestions: number;
}

export interface AdminDashboard {
  totalStudents: number;
  totalFaculty: number;
  totalExams: number;
  activeExams: number;
  totalResults: number;
  liveAttempts: number;
  recentResults: ExamResult[];
}

export interface FacultyDashboard {
  totalExams: number;
  totalMcqs: number;
  activeExams: number;
  recentExams: Exam[];
}

export type GetSubjectsParams = {
  programId?: number;
  semesterId?: number;
};

export type GetUsersParams = {
  role?: GetUsersRole;
};

export type GetUsersRole = (typeof GetUsersRole)[keyof typeof GetUsersRole];

export const GetUsersRole = {
  admin: "admin",
  faculty: "faculty",
  student: "student",
} as const;

export type GetAdminResultsParams = {
  examId?: number;
  programId?: number;
};

export type GetStudentListParams = {
  programId: number;
  yearId: number;
};