# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sehhat Foundation Online Exam System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/years:
    get:
      operationId: getYears
      tags: [admin]
      summary: List all academic years
      responses:
        "200":
          description: List of years
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Year"
    post:
      operationId: createYear
      tags: [admin]
      summary: Create academic year
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateYearRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Year"
  /admin/years/{id}:
    put:
      operationId: updateYear
      tags: [admin]
      summary: Update academic year
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateYearRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Year"
    delete:
      operationId: deleteYear
      tags: [admin]
      summary: Delete academic year
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/programs:
    get:
      operationId: getPrograms
      tags: [admin]
      summary: List all programs
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [admin]
      summary: Create program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /admin/programs/{id}:
    put:
      operationId: updateProgram
      tags: [admin]
      summary: Update program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [admin]
      summary: Delete program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/subjects:
    get:
      operationId: getSubjects
      tags: [admin]
      summary: List all subjects
      parameters:
        - name: programId
          in: query
          schema:
            type: integer
        - name: semesterId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [admin]
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /admin/subjects/{id}:
    put:
      operationId: updateSubject
      tags: [admin]
      summary: Update subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [admin]
      summary: Delete subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/users:
    get:
      operationId: getUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, faculty, student]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a user (admin/faculty/student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    put:
      operationId: updateUser
      tags: [admin]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/results:
    get:
      operationId: getAdminResults
      tags: [admin]
      summary: Get all exam results
      parameters:
        - name: examId
          in: query
          schema:
            type: integer
        - name: programId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamResult"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, faculty, student]
        programId:
          type: integer
          nullable: true
        yearId:
          type: integer
          nullable: true
        rollNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - role
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, faculty, student]
        programId:
          type: integer
          nullable: true
        yearId:
          type: integer
          nullable: true
        rollNumber:
          type: string
          nullable: true
      required:
        - username
        - name
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
          nullable: true
        programId:
          type: integer
          nullable: true
        yearId:
          type: integer
          nullable: true
        rollNumber:
          type: string
          nullable: true
...
    Year:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        value:
          type: integer
      required:
        - id
        - label
        - value
    CreateYearRequest:
      type: object
      properties:
        label:
          type: string
        value:
          type: integer
      required:
        - label
        - value
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: string
          nullable: true
      required:
        - id
        - name
        - code
    CreateProgramRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
          nullable: true
      required:
        - name
        - code
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        programId:
          type: integer
        semesterId:
          type: integer
        programName:
          type: string
          nullable: true
        semesterLabel:
          type: string
          nullable: true
      required:
        - id
        - name
        - code
        - programId
        - semesterId
    CreateSubjectRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        programId:
          type: integer
        semesterId:
          type: integer
      required:
        - name
        - code
        - programId
        - semesterId
...
    ExamResult:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
          nullable: true
        rollNumber:
          type: string
          nullable: true
        examId:
          type: integer
        examTitle:
          type: string
          nullable: true
        subjectName:
          type: string
          nullable: true
        programName:
          type: string
          nullable: true
        score:
          type: integer
        totalQuestions:
          type: integer
        percentage:
          type: number
          format: float
        submissionReason:
          type: string
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - examId
        - score
        - totalQuestions
        - percentage
        - submissionReason
        - submittedAt
...