# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sehhat Foundation Online Exam System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  programId?: number | null;
  yearId?: number | null;
  rollNumber?: string | null;
  createdAt: Date;
}