# lib/api-zod/src/generated/types/year.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sehhat Foundation Online Exam System API
 * OpenAPI spec version: 0.1.0
 */

export interface Year {
  id: number;
  label: string;
  value: number;
}